make content-input more simple
