+++ src/components/content-input.js
@@ -1,10 +1,7 @@
-import * as inputActions from 'actions/input';
-import * as operationActions from 'actions/operation';
-
 export default class ContentInputComponent {
-  constructor(target, store) {
+  constructor(target) {
     this.pressed = {};
-    this.store = store;
+    this.onKeyListeners = [];
 
     target.addEventListener('keypress', this.onKeyPress.bind(this));
     target.addEventListener('keydown', this.onKeyDown.bind(this));
@@ -12,6 +9,10 @@ export default class ContentInputComponent {
   }
 
   update() {
+  }
+
+  onKey(cb) {
+    this.onKeyListeners.push(cb);
   }
 
   onKeyPress(e) {
@@ -44,47 +45,20 @@ export default class ContentInputComponent {
     if (e.key === 'OS') {
       return;
     }
-    let keymaps = this.keymaps();
-    if (!keymaps) {
-      return;
-    }
-    this.store.dispatch(inputActions.keyPress(e.key, e.ctrlKey));
 
-    if (this.mapKeys(keymaps)) {
+    let stop = false;
+    for (let listener of this.onKeyListeners) {
+      stop = stop || listener(e.key, e.ctrlKey);
+    }
+    if (stop) {
       e.preventDefault();
       e.stopPropagation();
     }
-  }
-
-  mapKeys(keymaps) {
-    let input = this.store.getState().input;
-    let matched = Object.keys(keymaps).filter((keyStr) => {
-      return keyStr.startsWith(input.keys);
-    });
-    if (matched.length === 0) {
-      this.store.dispatch(inputActions.clearKeys());
-      return false;
-    } else if (matched.length > 1 ||
-      matched.length === 1 && input.keys !== matched[0]) {
-      return true;
-    }
-    let operation = keymaps[matched];
-    this.store.dispatch(operationActions.exec(operation));
-    this.store.dispatch(inputActions.clearKeys());
-    return true;
   }
 
   fromInput(e) {
     return e.target instanceof HTMLInputElement ||
       e.target instanceof HTMLTextAreaElement ||
       e.target instanceof HTMLSelectElement;
-  }
-
-  keymaps() {
-    let settings = this.store.getState().setting.settings;
-    if (!settings || !settings.json) {
-      return null;
-    }
-    return JSON.parse(settings.json).keymaps;
   }
 }
